Create new file
The module should model behavior similar to a memory with the ability to perform a read or write if given the address and the data to read or write. The memory is generic enough to allow other types of data to be accepted, as well as different widths of data and different ranges of addresses (although not specific ranges such as 32'h1000:32'h2000).