Create Dynamic_SGL_Letter_Stream.md
This is my first coding journal. It has working code so feel free to use it. I'm currently exploring the usefulness of Scatter-Gather Lists (SGLs) and what might be some typical applications as part of research related to work. SGLs are discussed in detail within the NVMe 1.4 specification which you can find here ->

https://nvmexpress.org/wp-content/uploads/NVM-Express-1_4b-2020.09.21-Ratified.pdf

The idea behind this code was to explore and experiment with using the SGL concept as a notepad for reads and writes. The rationale behind the idea was that if you populate an array of bytes treated as memory with random data then treat the entire array as a buffer containing only upper and lower cased ascii encoded characters then it should be possible to dynamically describe this "scattered buffer" with an SGL. I envisioned the alternating patterns of letters and non-letters as directly corresponding to alternating data block and bit bucket descriptors. I would read the entire array of bytes in linear time by making a single pass then create the appropriate descriptor for spans of letters and non-letter characters. Eventually I would create a list of the descriptors that ultimately would come to be one or more SGL segments.

I thought it was a good idea to start with that would help me sink my teeth into SGLs, but the results of my experiment were not encouraging even though I did learn a lot. An SGL is a data structure that describes a data buffer (which is either a source or a destination data buffer). A data block descriptor, for example, describes a block of data and a Bit Bucket Descriptor describes a Bit Bucket. Every descriptor requires 16 bytes of memory of it's own (if you're following the NVMe spec). Here are the results of the experiment summarized. You can also run the simulation yourself to see the results.

End of descriptor creation....
number_of_data_blocks =      10604
buffer_length_1 =      13429

So the number of data_blocks needed to describe all of the letters in the byte array is 10,604. This is because on average each descriptor only describes a data block that is approximately 1.2 bytes long which means that the spans of letters are either 1 or 2 letters long on average, but tend to be only one. Each descriptor takes up 16 bytes of memory, so the data block descriptors alone would consume 165 KB just to describe a little more than 13 KB worth of data and that doesn't even include the space occupied by the bit bucket descriptors! This is obviously not an efficient use of SGLs. I imagine anyone with experience working with SGLs would be laughing at me right now for not realizing it already, but I guess some things you have to find out for yourself - especially if nobody is around to tell you directly.