Create Intro_to_Programs.md
I'm currently learning to refine simulations through the use of the Program construct. My previous designs were ok, but programs are a nice way to separate the design from the testbench. There are plenty of specific situations where stimulus is generated or modified by a host environment or software in a test environment. For example, let's say the team is designing a smart home lock for the front door. The smart lock has physical hardware which needs to be designed and usually has some interfacing application or software whether it's a special app, suite of apps, cloud software, or just firmware with simple buttons (as in a digital combo).  The program constructs allows a modular approach to generating stimulus in response (to say key presses or user input) without needing to define an entire cloud application or software solution while also allowing the stimulus to be modified or replaced and all without affecting the hardware design aspect. The above examples come from the introduction to Programs in the SystemVerilog specifications.