Create Latch_Logic_Construct.sv
I had never used the always_latch construct before and recently needed to use latched values in a design. I was never even aware that there was such a thing. Ordinarily, I would have used an array and manipulated the values with continuous assignments or procedures or I might have used a memory type register, but this construct is much simpler. In Vivado, the netlist viewer shows only the two inputs (d and ck), the output (q), and names the cell q_reg which is labeled as a type of "RTL_LATCH".